Distill the game to movement: remove the energy system
The prototype is a rover, a hub, and the HTTP API. Everything that
supported energy — ports, capability records, the allocation system,
batteries, motors, panels, component replacement, grid commands — is
gone, along with its save fields. Speeds become floats; IMovable
declares the one gameplay capability.

- Server: routes reduced to /state, /objects/{id}, /objects/{id}/move
  and the session controls; request bodies are read explicitly in the
  route instead of BindAsync records
- Saves: version 4 only; saves carrying the removed energy objects are
  rejected instead of migrated — delete .space-automation/ to start a
  fresh expedition
- Remove the test project and the test recipe; .slnx becomes .sln
- README and the developer guide follow the smaller surface; ignore
  .vscode/

+++ backend/SpaceAutomation.Game/Rules.cs
@@ -3,11 +3,12 @@ namespace SpaceAutomation.Game;
 /// <summary>Shared gameplay validation helpers.</summary>
 public static class Rules
 {
-    public static bool Nonnegative(double value) => double.IsFinite(value) && value >= 0;
-    
+    public static bool Nonnegative(float value) => float.IsFinite(value) && value >= 0;
+
     public static void Require(bool condition, string message)
     {
-        if (!condition){
+        if (!condition)
+        {
             throw new InvalidDataException(message);
         }
     }
